Tweak user star ratings api call to support MW 1.27
Upgrading to MW 1.27 breaks this API call slightly.

I think this query can be optimized further, however this was the smallest set of
changes I could make which worked for both 1.24 and 1.27

Figured it out by directly manipulating the URL called by RW in the browser, and
finding one which produces the same results in both versions.